pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["stt*", "agent*", "memory*", "config*"]

[project]
name = "dolphin-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain[groq,ollama]>=1.2.15",
    "langgraph>=1.1.6",
    "openai-whisper>=20250625",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
    "scipy>=1.17.1",
    "sounddevice>=0.5.5",
    "textual>=8.2.3",
    "textual-dev>=1.8.0",
]
